+++ tests/phpunit/src/TestBase.php
@@ -209,7 +209,6 @@ abstract class TestBase extends TestCase {
     $this->clientProphecy->addOption('debug', Argument::type(OutputInterface::class));
     $this->clientServiceProphecy = $this->prophet->prophesize(ClientService::class);
     $this->clientServiceProphecy->getClient()->willReturn($this->clientProphecy->reveal());
-    $this->clientServiceProphecy->getConnector()->willReturn(new AccessTokenConnector(['access_token' => 'test', 'key' => NULL, 'secret' => NULL]));
     $this->logStreamManagerProphecy = $this->prophet->prophesize(LogstreamManager::class);
 
     $this->setIo($input, $output);
